+++ main/project/RapidCheck/target.h
@@ -14,11 +14,11 @@ public:
 	Target(){};
 	Target(Rect rect);
 	Target(Rect rect, MatND hist);
-	Target(Rect rect, MatND hist, double whiteRatio, double blackRatio);
+	Target(Rect rect, MatND hist, float whiteRatio, float blackRatio);
 	
 	Rect rect;
 	std::vector<Point> centerPositions;
-	double blackRatio, whiteRatio;
+	float blackRatio, whiteRatio;
 	bool found;
 	Point getCenterPoint();
